@@ -1,9 +1,11 @@
-import Vue from "vue";
 import App from "./App.vue";
-import vuetify from "./plugins/vuetify";
-import VueRouter from "vue-router";
+import {createRouter, createWebHistory} from "vue-router";
+import {createApp} from 'vue';
 
-import { firestorePlugin } from 'vuefire';
+import 'vuetify/styles';
+import { createVuetify } from 'vuetify';
+import * as components from 'vuetify/components';
+import * as directives from 'vuetify/directives';
 
 // pages
 
@@ -18,14 +20,13 @@ import ProfilePage from "@/pages/ProfilePage.vue";
 import AdminPage from "@/pages/AdminPage.vue";
 import RegisterPage from "@/pages/RegisterPage.vue";
 import RedirectRouter from "@/pages/RedirectRouter.vue";
-import AdminRoles from '@/pages/AdminRoles.vue'
-
-import VueYoutube from 'vue-youtube'
+import AdminRoles from '@/pages/AdminRoles.vue';
+import AdminStats from "@/pages/AdminStats.vue";
 
 import {auth} from './firebase';
 import { getUserPerms } from "./helpers";
-
-Vue.config.productionTip = false;
+import '@mdi/font/css/materialdesignicons.css';
+import '@/assets/override.css';
 
 const routes = [
   {
@@ -127,6 +128,38 @@ const routes = [
     },
   },
   {
+    path: "/admin/stats",
+    component: AdminStats,
+    meta: {
+      authRequired: true,
+      permsRequired: [[
+        "changeRolePerms",
+        "changeUserRole",
+        "editMyEvent",
+        "deleteMyEvent",
+        "acmAddEvent",
+        "acmEditEvent",
+        "acmDeleteEvent",
+        "icpcAddEvent",
+        "icpcEditEvent",
+        "icpcDeleteEvent",
+        "acmwAddEvent",
+        "acmwEditEvent",
+        "acmwDeleteEvent",
+        "broncosecAddEvent",
+        "broncosecEditEvent",
+        "broncosecDeleteEvent",
+        "otherAddEvent",
+        "otherEditEvent",
+        "otherDeleteEvent",
+        "viewAllResume",
+        "addProject",
+        "editProject",
+        "deleteProject"
+      ]]
+    }
+  },
+  {
     path: "/joinus",
     component: JoinUs,
   },
@@ -141,72 +174,63 @@ const routes = [
     }
   },
   {
-    path: "*",
+    path: '/:pathMatch(.*)*',
     redirect: "/",
-  },
+  }
 ];
 
-Vue.use(VueRouter);
-
-Vue.use(firestorePlugin);
- 
-Vue.use(VueYoutube)
-
-const router = new VueRouter({
-  base: "",
+const router = createRouter({
+  history: createWebHistory(),
   routes,
-  mode: "history",
   scrollBehavior() {
-    return { x: 0, y: 0 };
-  },
+    return { left: 0, top: 0 };
+  }
 });
 
-router.beforeEach( async (to, from, next) => {
+router.beforeEach( async (to, from) => {
   //Check if the page we are going to requires a user to be signed in or admin permissions
   const needsAuth = to.matched.some(record => record.meta.authRequired);
   const needsPerms = to.matched.find(record => record.meta.permsRequired)?.meta?.permsRequired;
   if (!needsAuth) {
-    next();
-    return;
+    return true;
   }
-
-  const user = auth.currentUser;
+  const user = await auth.currentUser;
+  let valid = false;
   if (user) {
-    let valid = true;
-
-    if(needsPerms && needsPerms.length > 0) {
+    valid = true;
+    if (needsPerms && needsPerms.length > 0) {
       const perms = await getUserPerms(user);
-      for(let permGroup of needsPerms) {
+      for (let permGroup of needsPerms) {
         let groupValid = false;
-        for(let perm of permGroup) {
-          if(perms[perm]) {
+        for (let perm of permGroup) {
+          if (perms[perm]) {
             groupValid = true;
           }
         }
-        if(!groupValid) {
+        if (!groupValid) {
           // console.log("FAILED for", permGroup, perms)
           valid = false;
         }
       }
     }
-    if(valid) {
-      next();
-      return;
-    }
+  }
+  if(valid) {
+    return true;
   }
 
   let path = "/redirect?uri="+encodeURIComponent(to.path);
   if(needsPerms && needsPerms.length > 0) {
     path += "&perms="+encodeURIComponent(needsPerms.map(row => row.join(",")).join(":"))
   }
-
-  next({
-    path: path,
-  });
+  return path;
 });
 
-new Vue({
-  vuetify,
-  router,
-  render: (h) => h(App),
-}).$mount("#app");
+const vuetify = createVuetify({
+  components,
+  directives
+});
+
+const app = createApp(App);
+app.use(router);
+app.use(vuetify);
+app.mount("#app")
